Working version of functions and types
The idea to define number of types, all of these types have schema and
are associated with yaml structure.

Function can work on a type and on a backend. This way it is possible to
implement new functions, enxtend yaml file functionality, add new
backends, chenge behaviour of functions, override functions and etc.