Add support for namespaces
Now there are two namespaces 'internal' and 'default'. If namespace is
not specified 'default' is always used. 'internal' namespace is
currently used for internal tables.

This solves issue with internal tables, because now there can be a
single manifest object with all the functions and types loaded, and more
than one manifest files loaded.