Move context.enter() to middleware
Previously streaming response returned from request handler function and
did iteration outside of handler function, this cuased code execute,
that expects to be withing context, but it was no longer in the context.

To fix that, move context block to middleware, where full
request/response handling is enclosed in the context.

Also added strict checks on Context class to make sure, that at the same
level one context variable can be set.

Also found, that probably there will be issues with context management
in async/thread pool invironments, this should be tested and fixed.