config.dump and config.read hardset argument
Moved config dumping code from cli to config as a method, to make it
easilly accessable for debugging.

Also chanded config.read(config=...) to config.read(hardeset=...). Hard
set takes highest priority and is usefull in situations, when for
example tests wan't to hardset 'env' to 'test', but for example .env
file has set 'env' to 'dev'. So with hardset, it is possible for tests
to tell, that no matter what, 'env' is always mast be set to 'test'.