Add <name> argument to `spinta config` command
You can use multiple options, example:

    > spinta config backends manifests
    Origin     Name                 Value
    ---------- -------------------- ----------------------------------------
    config     backends             default
    config       default            backend | dsn
    config         backend          spinta.backends.postgresql:PostgreSQL
    config         dsn              postgresql://admin:admin123@localhost:54321/spinta
    config     manifests            default
    config       default            backend | path
    config         backend          default
    config         path             manifest

Also added a header and switch Origin and Name positions. I think it is
more readable, when Name is closer to Value.

One more addition, if configuration value is a list it is displayer like
this `a | b | c`. I think this is more readable compared to python repr.