Add dataset and target properties for project
Projects should be able to choose which data source exatly they use. If
'dataset' is not specified, then it means, that project uses generic
model.

'target' is used for names used in project datase. And this project will
be used in export for project functionality, when exported data will
automatically use target names.