Fix NE operator on PostgreSQL backend
The fix is to use two joins if property belongs to a list, first join
checks if list has at ieast one value and second join checks if that
list does not contain given value.

Also found, that table keys are incorrectly written, but will fix that
in a next commit.