Prepare data with user defiend functions
The idea, that users should be able to define custom functions and thos
functions can be used in YAML files, for example:

    type: model
    name: country
    properties:
      code:
        type: string
        prepare: >-
          check(
            this, len(this) = 2,
            Country code must be 2 characters long!, got {given}.,
            given: len(this)
          )

After `code` property data is loaded, prepared and simple validation is
done, then user defined functions will be executed. In this case
`check` and `len` built-in user defined functions were used.

These functions are defined in `spinta.ufuncs` and look like this:

    @ufunc(Context, DataType)
    def check(context, dtype, this, value, message, **kwargs):
        if bool(value) is False:
            raise ValidationError(message.format(**kwargs))
        return this

    @ufunc(Context, DataType)
    def len(context, dtype, this):
        return len(this)

When prepareing values, each ufunc must return same or modified value
and it will be passed as `this` to the next ufunc.

Ufuncs are multiple dispatch functions, so it is possible to define
multiple ufuncs of the same name for different types.

Two arguments must always be passed to ufuncs, these are `context` and
`model` or `dtype`. These two argumens are not visible from manifest,
but all other arguments are the same as passed in yaml file.