Remove :external modifier
Original idea with :external was that it could work directly with on an
SQL database or a CSV file if :/external is specified. But also it could
read data from internal dtabase if :/external is not specified.

That means, one model can have two backends at the same time. But is not
true. One model at the same time can have only one backend. So there is
no need to have :external. Instead, one should used configuration, to
change backend.