Handle dataset model with no primary key
When primary key is not given, all model properties are used as primary
key. This can still generate duplicate primary keys, to fix that, better
option is to group all values and allso add count of each row, creating
a multiset. This would be more reliable, but probably I will add this
some time later.