Change repr of Expr
Decided to change repr of Expr into a nested function calls, in order to
better understand how callbacks will be called. While str of Expr shows
a human readable form.

Not sure if this is a good idea for Expr, but still it is nice to have
an option to unparse AST using nested function calls.