An experiment with multihosts
There was a requirement to host multiple Spinta instances on single
server. One option would be to configure separate Spinta instances, but
this whould require to update and configure two different Spinta
instances.

Another option, that I was playing with here, is to add support for
multiple hosts. This way, there can be one Spinta instance, but
configured to server multiple hosts, each host with a different
manifest.

Tried to add tests for this, but don't know yet if I will work on this,
or just go with separate installs. Adding multiple hosts support would
be nice, but might require too many time to implement.