Fix streaming issue and upgrade Starlette
After a lot of debugging, I dug out, that streaming wasn't working
because of BaseHTTPMiddleware, which it seems collects all the stream
into memory and then returns it all at once with the response.

That means, if you stream a lot of data, request will not give any
answer until all data is collected into memory. This can take time and
can result in read timeout or can simply run out of memory.

Streaming was the main reason, why I chose Starlette, and one and most
important thing didn't worked. Not good. But at least I found how to fix
it.

Related issues:

https://github.com/encode/starlette/issues/1012
https://github.com/encode/starlette/issues/472