Backends in manifest and inspect command
Many changes, related with two tickets:

- https://gitlab.com/atviriduomenys/spinta/-/issues/50
- https://gitlab.com/atviriduomenys/spinta/-/issues/49

But implementation for both tickets is not complete.

Now it is possible to define backend in manifest, but I didn't tested if
defined backend in manifest is really used.

Also a basic version of inspect command for sql backend is implemented,
but it does not support inspecting primary and foreign keys and probably
many other things.

Inspect part is based on frictionless describe functionality, so with
that I will get all formats supported by frictionless.