Inspect resource given as command line argument
In order to make this work I had to dynamically generate a manifest and
inject given resource into it. In order to generate a manifest, decided
to reuse a functionality, that allows to generate manifest from an ascii
table. In order to do that, had to move ascii table loading functions
from testing into spinta.manifests.tabular.helpers. But stumbled on
multiple errors there.

Alse there were issues with S3 backend, which is hardcoded into test
config and is available globally and requires s3 fixture. For now, just
rewrote backends in tests. But this needs to be fixed.