Manifest configuration helper
Created a helper which is responsible for configuring manifest. Now many
command can operate just from command line, without an existing manifest
or with one or more given manifests via command line or with a manifest
generated from given resources.

All this functionality is needed for multiple commands, so decided to
move it into helpers. Currently this is used only with `spinta run`
command.

Also added `Backend.origin` in order to know when backend was defined
inline, when it was defined in a manifest or in a config. This helps to
render manifests tables from `Manifest` instances more precisely.

And some effort went into `pp` debug tool.