Remove query and improve file access log
I found that query logging is not consistent, because UrlParams does not
provide full query, some parts are moved to the attributes.

So decided to remove it completely, query is available in url anyway.
But for writes, _where part should be logged, but it is not logged right
now, added a TODO mark for this.

Also refactored FileAccessLog a bit, added tests for stdin, stdout and
null cases.