text type initial support
It is possible to use text type in manifest files, write and read
to/from database. Added tests for html format, but could not easilly get
it show `name@lang`, now it shows `name.lang`.

Didn't tested it with on changelog, didn't tested how patch and
changelog works.

Didn't tested if it is able to read and push text type from external
sources.