WSDL support (WIP)
Started with just WSDL, that went surprisingly well, but then decided to
refactor existing SQL schema reader and move out of frictionless data
into sqlalchemy directly, this was also done.

But then decided to refactor merging of schemas out of specific manifest
implementation to an upper level in order to make this available to all
manifests, but didn't finished it yet.